Implement four nested classes to calculate each part of the MPD triangle
Inside ThreadedMinimumPairwiseDistance, there are four nest classes: LowerLeftMPD, BottomRightMPD, TopRightMPD and CenterMPD. Each nested class calculates each part of the MPD triangle (as described in the README) and compares the local result of each class to the global result. If the local result is smaller than the global result, the global result is then the local result (aka the minimum pairwise distance). All tests pass.